refactor: improve build system and refine Rust driver compatibility
Update the build configuration and refine the integration between the
MINIX kernel and Rust-based drivers.

Key changes include:
- Build System:
  - Enable CXX support in the top-level CMake configuration.
  - Improve cross-compilation reliability by explicitly passing the
    target triple to Clang via CMAKE_C/CXX/ASM_FLAGS.
  - Add support for `do_capctl.c` and `mac_hooks.c` in the kernel build.
- Kernel:
  - Refine x86_64 context switching and IRQ thread entry logic.
  - Add `kaslr_virt_offset_slot` to user-mapped data for KASLR support.
  - Fix memory mapping logic in `protect.c`.
- Rust Drivers & Libraries:
  - Transition many crates from strict `#![no_std]` to conditional
    `#![cfg_attr(target_os = "minix", no_std)]` to improve host-side
    testing and development.
  - Refine `ext4-core` with a `std` feature for `mkfs` functionality.
  - Fix various driver-specific bugs, including NVMe register layout
    corrections, Bluetooth gamepad profile detection order, and
    XHCI DMA allocation stubs for unit testing.
  - Correct several test assertions related to memory layout and
    hardware register offsets.

+++ rust/minix-ahci/src/registers.rs
@@ -435,12 +435,12 @@ mod ffi_tests {
     fn ahci_test_reg_offsets_are_correct() {
         unsafe {
             // HBA registers (byte offsets = index * 4)
-            assert_eq!(ahci_test_reg_offset(0), hba::CAP * 4);   // HBA_CAP
-            assert_eq!(ahci_test_reg_offset(1), hba::GHC * 4);   // HBA_GHC
-            assert_eq!(ahci_test_reg_offset(2), hba::IS * 4);    // HBA_IS
-            assert_eq!(ahci_test_reg_offset(3), hba::PI * 4);    // HBA_PI
-            assert_eq!(ahci_test_reg_offset(12), port::IS * 4);  // PORT_IS
-            assert_eq!(ahci_test_reg_offset(14), port::CMD * 4); // PORT_CMD
+            assert_eq!(ahci_test_reg_offset(0), (hba::CAP * 4) as u32);   // HBA_CAP
+            assert_eq!(ahci_test_reg_offset(1), (hba::GHC * 4) as u32);   // HBA_GHC
+            assert_eq!(ahci_test_reg_offset(2), (hba::IS * 4) as u32);    // HBA_IS
+            assert_eq!(ahci_test_reg_offset(3), (hba::PI * 4) as u32);    // HBA_PI
+            assert_eq!(ahci_test_reg_offset(12), (port::IS * 4) as u32);  // PORT_IS
+            assert_eq!(ahci_test_reg_offset(14), (port::CMD * 4) as u32); // PORT_CMD
 
             // Unknown ID
             assert_eq!(ahci_test_reg_offset(99), 0xFFFFFFFF);
